requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
certifi==2024.7.4
    # via
    #   requests
    #   roboflow
chardet==4.0.0
    # via roboflow
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via
    #   matplotlib
    #   roboflow
easyocr==1.7.1
    # via -r requirements.in
filelock==3.15.4
    # via torch
filetype==1.2.0
    # via roboflow
fonttools==4.53.1
    # via matplotlib
fsspec==2024.6.1
    # via torch
idna==3.7
    # via
    #   requests
    #   roboflow
imageio==2.34.2
    # via scikit-image
intel-openmp==2021.4.0
    # via mkl
jinja2==3.1.4
    # via torch
kiwisolver==1.4.5
    # via
    #   matplotlib
    #   roboflow
lazy-loader==0.4
    # via scikit-image
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.1
    # via roboflow
mkl==2021.4.0
    # via torch
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via
    #   scikit-image
    #   torch
ninja==1.11.1.1
    # via easyocr
numpy==2.0.1
    # via
    #   -r requirements.in
    #   contourpy
    #   easyocr
    #   imageio
    #   matplotlib
    #   opencv-python-headless
    #   roboflow
    #   scikit-image
    #   scipy
    #   shapely
    #   tifffile
    #   torchvision
opencv-python-headless==4.10.0.84
    # via
    #   easyocr
    #   roboflow
packaging==24.1
    # via
    #   lazy-loader
    #   matplotlib
    #   pytesseract
    #   scikit-image
pillow==10.4.0
    # via
    #   -r requirements.in
    #   easyocr
    #   imageio
    #   matplotlib
    #   pytesseract
    #   roboflow
    #   scikit-image
    #   torchvision
pyclipper==1.3.0.post5
    # via easyocr
pyparsing==3.1.2
    # via matplotlib
pytesseract==0.3.10
    # via -r requirements.in
python-bidi==0.6.0
    # via easyocr
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   roboflow
python-dotenv==1.0.1
    # via roboflow
pyyaml==6.0.1
    # via
    #   easyocr
    #   roboflow
requests==2.32.3
    # via
    #   requests-toolbelt
    #   roboflow
requests-toolbelt==1.0.0
    # via roboflow
roboflow==1.1.37
    # via -r requirements.in
scikit-image==0.24.0
    # via easyocr
scipy==1.14.0
    # via
    #   easyocr
    #   scikit-image
shapely==2.0.5
    # via easyocr
six==1.16.0
    # via
    #   python-dateutil
    #   roboflow
sympy==1.13.1
    # via torch
tbb==2021.13.0
    # via mkl
tifffile==2024.7.24
    # via scikit-image
torch==2.3.1
    # via
    #   easyocr
    #   torchvision
torchvision==0.18.1
    # via easyocr
tqdm==4.66.4
    # via roboflow
typing-extensions==4.12.2
    # via torch
urllib3==2.2.2
    # via
    #   requests
    #   roboflow
